Create crystal_frequency_generator.rs
This is a simple example of a struct CrystalFrequencyGenerator that has two fields frequency and crystal_type and two methods generate_sine_wave and generate_frequency_sweep. The new function is a constructor that takes a crystal type as a string and returns a new instance of CrystalFrequencyGenerator with the frequency field set based on the type of crystal passed. The generate_sine_wave function returns a vector of 1000 samples of a sine wave with the frequency set to the frequency of the crystal and generate_frequency_sweep returns a vector of frequency values that linearly sweep from start_freq to stop_freq.