Update quantum_teleporter.rs
In this version, the entangled_pairs field has been changed from a vector of tuples to a HashMap where the keys are qubit names and the values are the names of their entangled partners. This allows for faster lookups when searching for entangled pairs.

Note that when adding entangled pairs, we insert both (qubit1, qubit2) and (qubit2, qubit1) into the entangled_pairs HashMap. This is because entanglement is a symmetric relationship.