Create main.rs
In this example, the main function creates an instance of the TeleportationCoordinator and UserInterface structs, starts the teleportation process using the start method of the TeleportationCoordinator, and then displays the results using the display_results method of the UserInterface. These methods and structs are defined in the corresponding .rs files in the internal directory.